Fixed signage problems with texture_i typedef; fixed functions for text due to unused parameter
git-svn-id: svn://svn.code.sf.net/p/freeband2/code@58 611cfc87-9d2d-4324-8440-b07dc6783d85

+++ src/input/screenDifficulty.c
@@ -7,7 +7,7 @@
 #include "screenDifficulty.h"
 
 GLvoid input_screenDifficulty(GLuint direction) {
-  GLuint i;
+  GLushort i;
   
   switch (direction) {
     case DOWN:
